refactor/simplify code
